fix: update plugin configs and reorganize files
