+++ packages/gatsby-theme-aio/algolia/algolia-search-settings.js
