+++ packages/gatsby-theme-aio/algolia/helpers/get-content-from-cache.js
