+++ packages/gatsby-theme-aio/algolia/helpers/get-iframe-content.js
