+++ packages/gatsby-theme-aio/algolia/helpers/get-imported-content.js
