+++ packages/gatsby-theme-aio/algolia/helpers/get-openapi-content.js
