+++ packages/gatsby-theme-aio/algolia/helpers/get-products-indexes.js
