+++ packages/gatsby-theme-aio/algolia/helpers/html-parser.js
