+++ packages/gatsby-theme-aio/algolia/helpers/parse-html.js
