+++ packages/gatsby-theme-aio/algolia/helpers/parse-markdown.js
